feat: authenticated access for authors

+++ backend/src/services/posts.js
@@ -1,21 +1,22 @@
 import { Post } from '../db/models/post.js'
+import { User } from '../db/models/user.js'
 
-export async function createPost({ title, author, contents, tags }) {
-  const post = new Post({ title, author, contents, tags })
+export async function createPost(userId, { title, contents, tags }) {
+  const post = new Post({ title, author: userId, contents, tags })
   return await post.save()
 }
 
-export async function getPost(id) {
+export async function getPostById(id) {
   return await Post.findById(id)
 }
 
-export async function updatePost(id, { title, author, contents, tags }) {
+export async function updatePost(userId, id, { title, contents, tags }) {
   return await Post.findOneAndUpdate(
-    { _id: id },
+    { _id: id, author: userId },
     {
       $set: {
         title,
-        author,
+        author: userId,
         contents,
         tags,
       },
@@ -24,16 +25,18 @@ export async function updatePost(id, { title, author, contents, tags }) {
   )
 }
 
-export async function deletePost(id) {
-  return await Post.deleteOne({ _id: id })
+export async function deletePost(userId, id) {
+  return await Post.deleteOne({ _id: id, author: userId })
 }
 
 export async function listAllPosts(options) {
   return await listPosts({}, options)
 }
 
-export async function listPostsByAuthor(author, options) {
-  return await listPosts({ author }, options)
+export async function listPostsByAuthor(authorUsername, options) {
+  const user = await User.findOne({ username: authorUsername })
+  if (!user) return []
+  return await listPosts({ author: user._id }, options)
 }
 
 export async function listPostsByTag(tags, options) {
